# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  requests.PartnershipDeclineRequest:
    properties:
      reason:
        maxLength: 500
        type: string
    type: object
...
  requests.PartnershipRequestCreate:
    properties:
      message:
        maxLength: 2000
        minLength: 10
        type: string
      opportunity_id:
        type: string
      requested_types:
        items:
          type: string
        maxItems: 10
        minItems: 1
        type: array
    required:
    - message
    - requested_types
    type: object
...
  responses.BrandSummary:
    properties:
      business_name:
        type: string
      id:
        type: string
      industry:
        type: string
      location:
        type: string
      logo_url:
        type: string
    type: object
...
  responses.ErrorDetail:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  responses.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/responses.ErrorDetail'
      request_id:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
...
  responses.PartnershipDashboardResponse:
    properties:
      accepted_count:
        type: integer
      active_count:
        type: integer
      cancelled_count:
        type: integer
      completed_count:
        type: integer
      declined_count:
        type: integer
      pending_count:
        type: integer
      total_count:
        type: integer
    type: object
  responses.PartnershipEventRef:
    properties:
      currency:
        type: string
      end_at:
        type: string
      id:
        type: string
      slug:
        type: string
      start_at:
        type: string
      thumbnail_url:
        type: string
      title:
        type: string
    type: object
...
  responses.PartnershipRequestResponse:
    properties:
      accepted_at:
        type: string
      activated_at:
        type: string
      brand:
        allOf:
        - $ref: '#/definitions/responses.BrandSummary'
        description: Embedded summary info for cards
      brand_profile_id:
        type: string
      brand_user_id:
        type: string
      cancelled_at:
        type: string
      completed_at:
        type: string
      created_at:
        type: string
      declined_at:
        type: string
      event:
        $ref: '#/definitions/responses.PartnershipEventRef'
      event_id:
        type: string
      id:
        type: string
      message:
        type: string
      opportunity_id:
        type: string
      organizer_id:
        type: string
      reason:
        type: string
      requested_types:
        items:
          type: string
        type: array
      status:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@1kulture.com
    name: API Support
    url: https://1kulture.com/support
  description: Enterprise Event Management System API with multi-tenant SaaS architecture
  license:
    name: Proprietary
    url: https://1kulture.com/license
  termsOfService: https://1kulture.com/terms
  title: 1Kulture API
  version: "1.0"
paths:
...
  /admin/partnerships:
    get:
      consumes:
      - application/json
      description: Admin only.
      parameters:
      - description: Status
        in: query
        name: status
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Partnerships
          schema:
            items:
              $ref: '#/definitions/responses.PartnershipRequestResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all partnerships (admin)
      tags:
      - admin-partnerships
...
  /brands/partnerships:
    get:
      consumes:
      - application/json
      description: Requests made by the authenticated brand.
      parameters:
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Requests
          schema:
            items:
              $ref: '#/definitions/responses.PartnershipRequestResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List my partnership requests (brand)
      tags:
      - partnerships
...
  /events/{id}/partnership-requests:
    get:
      consumes:
      - application/json
      description: Organizer/co-organizer only.
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Requests
          schema:
            items:
              $ref: '#/definitions/responses.PartnershipRequestResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List partnership requests for an event (organizer)
      tags:
      - partnerships
    post:
      consumes:
      - application/json
      description: Brand requests to partner with an event.
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      - description: Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.PartnershipRequestCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.PartnershipRequestResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "409":
          description: Already requested
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Request a partnership
      tags:
      - partnerships
...
  /partnerships/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Partnership Request ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Partnership
          schema:
            $ref: '#/definitions/responses.PartnershipRequestResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a partnership
      tags:
      - partnerships
  /partnerships/{id}/activate:
    post:
      consumes:
      - application/json
      description: Organizer/co-organizer only.
      parameters:
      - description: Partnership Request ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Activated
          schema:
            $ref: '#/definitions/responses.PartnershipRequestResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Mark a partnership as active
      tags:
      - partnerships
...
  /partnerships/{id}/cancel:
    post:
      consumes:
      - application/json
      parameters:
      - description: Partnership Request ID
        in: path
        name: id
        required: true
        type: string
      - description: Reason
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.PartnershipDeclineRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Cancelled
          schema:
            $ref: '#/definitions/responses.PartnershipRequestResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cancel a partnership request (brand)
      tags:
      - partnerships
  /partnerships/{id}/complete:
    post:
      consumes:
      - application/json
      description: Organizer/co-organizer only.
      parameters:
      - description: Partnership Request ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Completed
          schema:
            $ref: '#/definitions/responses.PartnershipRequestResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Mark a partnership as completed
      tags:
      - partnerships
  /partnerships/{id}/decide:
    post:
      consumes:
      - application/json
      description: Organizer/co-organizer only.
      parameters:
      - description: Partnership Request ID
        in: path
        name: id
        required: true
        type: string
      - description: Accept or decline
        in: query
        name: accept
        required: true
        type: boolean
      - description: Reason (for decline)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.PartnershipDeclineRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Decided
          schema:
            $ref: '#/definitions/responses.PartnershipRequestResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Approve or decline a partnership request (organizer)
      tags:
      - partnerships
...
  /partnerships/dashboard:
    get:
      consumes:
      - application/json
      description: 'Organizer dashboard: counts by status. Pass event_id to scope
        to one event.'
      parameters:
      - description: Optional Event ID to scope
        in: query
        name: event_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Dashboard
          schema:
            $ref: '#/definitions/responses.PartnershipDashboardResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Partnership dashboard
      tags:
      - partnerships
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"